# docs/api/openapi.yaml
components:
  schemas:
    ApplicantRequest:
      description: |
        **Restricted personal data.** Supplied on creation and never returned by any endpoint. It is stored encrypted, is not written to any log, and only the derived pseudonym crosses a service boundary.
      properties:
        dateOfBirth:
          format: date
          type: string
        emailAddress:
          format: email
          maxLength: 320
          minLength: 0
          type: string
        familyName:
          maxLength: 200
          minLength: 0
          type: string
        givenName:
          maxLength: 200
          minLength: 0
          type: string
        residenceCountry:
          minLength: 1
          pattern: "^[A-Za-z]{2}$"
          type: string
      required:
      - dateOfBirth
      - emailAddress
      - familyName
      - givenName
      - residenceCountry
      type: object
    ApplicationResponse:
      properties:
        amountMinorUnits:
          description: "The requested principal in the currency's MINOR units - cents, not euros. Money is never a floating-point number here."
          example: 1500000
          format: int64
          type: integer
        applicantInitials:
          description: "Initials only. The API never returns an applicant's name, not even to the client that supplied it."
          example: T.A.
          type: string
        applicantReference:
          description: |
            An irreversible HMAC pseudonym, stable for the same applicant across applications.

            **Pseudonymised, not anonymised**: the same applicant always produces the same value, so it is linkable and remains personal data under GDPR.
          example: APR-fa05675c49b42435
          pattern: "^APR-[0-9a-f]{16}$"
          type: string
        applicationId:
          description: "Opaque, server-generated."
          format: uuid
          type: string
        createdAt:
          format: date-time
          type: string
        currency:
          type: string
        decision:
          $ref: "#/components/schemas/Decision"
          description: Null until a decision has been recorded.
        productCode:
          type: string
        purpose:
          type: string
        residenceCountry:
          type: string
        status:
          description: The application's position in its lifecycle.
          enum:
          - DRAFT
          - SUBMITTED
          - VALIDATING
          - CHECKS_IN_PROGRESS
          - MANUAL_REVIEW
          - APPROVED
          - REJECTED
          - FAILED
          - CANCELLED
          type: string
        submittedAt:
          format: date-time
          type: string
        termMonths:
          format: int32
          type: integer
        updatedAt:
          format: date-time
          type: string
        version:
          description: Optimistic-locking version.
          format: int64
          type: integer
      type: object
    ApplicationStatusResponse:
      properties:
        applicationId:
          format: uuid
          type: string
        decisionType:
          type: string
        reasonCode:
          type: string
        status:
          enum:
          - DRAFT
          - SUBMITTED
          - VALIDATING
          - CHECKS_IN_PROGRESS
          - MANUAL_REVIEW
          - APPROVED
          - REJECTED
          - FAILED
          - CANCELLED
          type: string
        terminal:
          description: Whether the status can still change. **Stop polling when this is true.**
          type: boolean
        updatedAt:
          format: date-time
          type: string
        version:
          format: int64
          type: integer
      type: object
...
    CreateApplicationRequest:
      properties:
        applicant:
          $ref: "#/components/schemas/ApplicantRequest"
        loan:
          $ref: "#/components/schemas/LoanRequestPayload"
      required:
      - applicant
      - loan
      type: object
    Decision:
      description: "The recorded outcome, once there is one."
      properties:
        decidedBy:
          description: "The deciding reviewer's pseudonymous token subject, or null when the decision was automatic. Never a name."
          type: string
        reasonCode:
          description: "A stable code, never prose. Free text about why a named individual was refused credit is exactly what must not travel through events into the audit store."
          example: CREDIT_SCORE_BELOW_THRESHOLD
          type: string
        type:
          description: |
            `MANUAL_REVIEW` means the platform declined to decide. That is not the same as deciding against the applicant, and the record has to be able to tell the two apart later.
          enum:
          - APPROVED
          - REJECTED
          - MANUAL_REVIEW
          type: string
      type: object
...
    LoanRequestPayload:
      properties:
        amountMinorUnits:
          description: "The requested principal in the currency's MINOR units - cents, not euros. Money is never a floating-point number here."
          example: 1500000
          format: int64
          minimum: 0
          type: integer
        currency:
          minLength: 1
          pattern: "^[A-Z]{3}$"
          type: string
        declaredAnnualIncomeMinorUnits:
          description: "Declared gross annual income, in minor units."
          example: 6000000
          format: int64
          minimum: 0
          type: integer
        productCode:
          description: Internal product identifier.
          example: PERSONAL-LOAN-STANDARD
          minLength: 1
          type: string
        purpose:
          enum:
          - HOME_IMPROVEMENT
          - DEBT_CONSOLIDATION
          - VEHICLE
          - EDUCATION
          - MEDICAL
          - BUSINESS
          - OTHER
          minLength: 1
          type: string
        termMonths:
          format: int32
          minimum: 1
          type: integer
      required:
      - amountMinorUnits
      - currency
      - declaredAnnualIncomeMinorUnits
      - productCode
      - purpose
      - termMonths
      type: object
    MissingDocumentsProblem:
      description: Returned when an application cannot be submitted because a mandatory document has not been uploaded and accepted.
      properties:
        errorCode:
          example: MANDATORY_DOCUMENTS_MISSING
          type: string
        missingDocumentTypes:
          description: "The document CATEGORIES that are missing. Never a document's contents, and never a filename."
          enum:
          - PROOF_OF_IDENTITY
          - PROOF_OF_INCOME
          items:
            type: string
          type: array
      type: object
...
    Problem:
      description: |
        An RFC 9457 Problem Detail.

        It never contains a stack trace, a SQL fragment, a table name, an internal identifier, or an exception message. Validation errors name the **field** and never echo its value, because the value is frequently the personal data the platform exists to protect.

        Branch on `errorCode`, never on `detail`.
      properties:
        correlationId:
          description: "Ties this response to the platform's logs. Quote it when reporting a problem: it is deliberately the only link, because the body carries nothing else."
          type: string
        detail:
          description: Prose for a human. May be reworded between releases.
          type: string
        errorCode:
          description: Stable and machine-readable. This is what a client branches on.
          enum:
          - REQUEST_VALIDATION_FAILED
          - MALFORMED_REQUEST_BODY
          - INVALID_REQUEST_VALUE
          - INVALID_PATH_PARAMETER
          - ACCESS_DENIED
          - APPLICATION_NOT_FOUND
          - APPLICATION_NOT_EDITABLE
          - DOCUMENT_NOT_FOUND
          - ILLEGAL_STATE_TRANSITION
          - ILLEGAL_DOCUMENT_STATE
          - MANDATORY_DOCUMENTS_MISSING
          - IDEMPOTENCY_KEY_REUSED
          - IDEMPOTENCY_KEY_IN_USE
          - CONCURRENT_MODIFICATION
          - CONCURRENT_DOCUMENT_UPDATE
          - OBJECT_STORAGE_UNAVAILABLE
          - SCANNER_UNAVAILABLE
          - INTERNAL_ERROR
          type: string
        instance:
          example: /v1/applications/018f3a9c-6b21-7c4e-9d55-2a1b8e7f0c33/submit
          type: string
        status:
          example: 409
          format: int32
          type: integer
        timestamp:
          format: date-time
          type: string
        title:
          example: Conflict
          type: string
        type:
          example: about:blank
          type: string
      type: object
...
    StateTransitionProblem:
      description: Returned when the requested operation is not valid for the application's current state.
      properties:
        attemptedStatus:
          description: The state the request would have moved it to.
          type: string
        currentStatus:
          description: The state the application is actually in.
          type: string
        errorCode:
          example: ILLEGAL_STATE_TRANSITION
          type: string
      type: object
...
  securitySchemes:
    oauth2:
      description: |
        An OAuth2 bearer token, validated at API Gateway **and again by each service**. The second validation is what protects the services if anything is ever deployed with a route that bypasses the gateway.

        Partners additionally present an API key. The key is what a usage plan throttles on — a capacity identity, never an authentication one — so a key alone grants nothing.
      flows:
        clientCredentials:
          scopes:
            applications:read: Read applications and their status
            applications:write: "Create, update and submit applications"
            documents:read: List documents on an application
            documents:write: Request upload slots and confirm uploads
            manual-review:decide: Record a reviewer's decision
            manual-review:read: Read the manual review queue
          tokenUrl: https://auth.example.com/oauth2/token
      type: oauth2
info:
  contact:
    name: Repository
    url: https://example.com/aws-loan-origination-platform
  description: |
    ## What this is

    A **reference implementation**. It has never been deployed, it has never processed real data, and every external check behind it — KYC, AML, fraud, credit scoring, malware scanning — is a **simulation** that contacts no real provider.

    ## Two services, one API surface

    These paths are served by two independently deployed services behind a single API Gateway. `/v1/applications/{applicationId}/documents/**` belongs to the document service; everything else belongs to the application service. Callers should not depend on that split — it is recorded because it explains why a document can be accepted a moment before submission will accept it.

    ## Idempotency

    `POST /v1/applications` and `POST .../submit` accept an `Idempotency-Key` header, and submission is the one that most needs it: a client that times out and retries must not submit twice.

    | Situation | Result |
    |---|---|
    | Same key, same request | The original response, with `Idempotent-Replay: true` |
    | Same key, different request | `409` with `IDEMPOTENCY_KEY_REUSED` |
    | Same key, original still in flight | `409` with `IDEMPOTENCY_KEY_IN_USE` |

...

    ## Errors

...

    Branch on `errorCode`, never on `detail`, which is prose and may be reworded.

    ## Personal data

...

...
  license:
    identifier: Apache-2.0
    name: Apache-2.0
...
  version: 1.0.0
openapi: 3.1.0
paths:
  /v1/applications:
    post:
      description: |
        Creates the application in `DRAFT`. Nothing is assessed until it is submitted, and it cannot be submitted until every mandatory document has been uploaded and accepted.
      operationId: createDraft
      parameters:
      - description: "A client-generated key, unique per logical operation. Reuse it when retrying the SAME request; never reuse it for a different one."
        example: 3f8c1b6e-9a2d-4c7f-8f10-2b5d9a1c4e77
        in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationRequest"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApplicationResponse"
          description: "Created, or replayed. A replay carries `Idempotent-Replay: true` and is the original response, not a new application."
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Problem"
          description: The request failed validation.
        "401":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Problem"
          description: "No token, or a token that is expired or minted for another audience."
        "403":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Problem"
          description: The token lacks `applications:write`.
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The `Idempotency-Key` was already used for a different request, or the original is still in flight."
      summary: Create a draft application
      tags:
      - Applications
  /v1/applications/{applicationId}:
    get:
      description: |
        The full representation. It identifies the applicant by pseudonym and initials, never by the values that were supplied.
      operationId: getApplication
      parameters:
      - in: path
        name: applicationId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApplicationResponse"
          description: The application.
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Problem"
          description: "No such application. Deliberately indistinguishable from \"exists but is not yours\": telling the two apart lets a caller enumerate which identifiers exist."
      summary: Retrieve an application
      tags:
      - Applications
...
  /v1/applications/{applicationId}/draft:
    put:
      description: |
        Refused once the application has been submitted. Changing the amount after submission would mean the assessment ran against figures nobody submitted.
      operationId: updateDraft
      parameters:
      - in: path
        name: applicationId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApplicationResponse"
          description: The updated draft.
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/StateTransitionProblem"
          description: The application is no longer a draft. The body names the current and attempted states.
      summary: Replace the contents of a draft
      tags:
      - Applications
  /v1/applications/{applicationId}/status:
    get:
      description: |
        Deliberately small and separate from the full representation. Status polling is the highest-volume read on the platform, and a client waiting for a decision should not transfer the whole application on every poll.

        Stop polling when `terminal` is true.
      operationId: getStatus
      parameters:
      - in: path
        name: applicationId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApplicationStatusResponse"
          description: The current status.
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Problem"
          description: No such application.
      summary: The current status
      tags:
      - Applications
  /v1/applications/{applicationId}/submit:
    post:
      description: |
        Starts the assessment. Returns **202 Accepted**, not 200: the response says the request was accepted, not that a decision was reached. Poll `/status` for the outcome.

        There is no request body. The path identifier is what the idempotency fingerprint is taken over, so reusing a key against a different application is a conflict rather than a silent replay of the first application's result — which would tell a client its second application had been submitted when it had not.
      operationId: submit
      parameters:
      - in: path
        name: applicationId
        required: true
        schema:
          type: string
      - description: See the create operation.
        in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
      responses:
        "202":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApplicationResponse"
          description: "Accepted for assessment, or replayed."
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Problem"
          description: No such application.
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Problem"
          description: Idempotency key conflict.
        "422":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MissingDocumentsProblem"
          description: "A mandatory document has not been uploaded and accepted. The body names the missing CATEGORIES, never a document's contents."
      summary: Submit an application for assessment
      tags:
      - Applications
...
security:
- oauth2: []
servers:
...